Create feedback form (#9)
Created a feedback from with some basic inital questions. Also implemented a QR scanner that can scan a QR code and send the data back to the form.

Note: I used the "Next" version of expo-camera, which is like a preview/beta version of an upcoming update to the current library. It's supposedly stable, but at one point, my Expo Go app kept crashing without an error whenever I opened the camera. I had to reinstall the Expo Go app to get it back working. I hope it's just a random fluke, but if it happens again, we may need to revert to the current version of expo-camera.

---------

Co-authored-by: samuelparada6 <samuelparada6@vt.edu>