Create user database (#14)
Uses device storage for favorite routes. I know we initially thought of using a DB, but that's pretty inefficient for now. If in the future we find we need a more robust storage system, we can migrate then. For now, device storage should be sufficient. I added methods for adding/fetching/removing favorite routes and stops. JavaDocs are available to help out with VSCode intelliensene and knowing how to use them.