Remove Alerts tab (#18)
I replaced the Alerts Tab with a Favorites Tab, as Grace and I discussed. I updated the UI buttons on the top right section of the Maps view. I also added the BottomSheet (swipe-bar) on the Routes and Favorites tabs. The style of them are rough and will be enhanced soon. And rearranged the tab order of the app.