Add connected routes (#19)
Give it a start and stop location (coordinates or string address) and it'll find connected routes. It'll return an array of objects, where each object is a leg/route for the whole trip. The properties it has are listed in the JSDoc comment. Not sure what data we need, but I returned a few that I thought might be useful. As the frontend needs more data, we can add more.