Fix map (#27)
Worked on getting the Map component to be synchronized across all tabs by lifting the state from the Map.js file to the App.js file to then pass down the state props to all the tabs to have access to. This method is buggy and laggy, so this is not a definite solution, but temporary. Will be looking into a more optimal solution in a couple of days. I also fixed a UI issue with the 'more options' button on Navigate.