Connect Routes (#39)
Reworked backend to provide vital information to display to the user on the frontend side. Connected the backend to the frontend in the Navigate tab. Added styling and color and all the creativity to finding a route from start to end destinations. Added another stack screen to navigate to once a route was configured. Displayed a colored route based on bus color from one location to another with instructions in a BottomSheet.